refactor(state): extract Telegram topic cluster into SessionTelegramTopicsMixin
SessionDB is 6,868 lines and 153 methods. It is already a mixin composition:

    class SessionDB(SessionSearchMixin, SessionSchemaMixin, SessionPortabilityMixin)

This adds the fourth, following hermes_state_search.py, hermes_state_schema.py
and hermes_state_portability.py exactly.

The cluster is 492 contiguous lines (7786-8287) of Telegram forum-topic
support: the opt-in schema migration, enable/disable/query of topic mode, and
CRUD over the chat-topic-to-session bindings it creates. It is the only
platform-specific code in SessionDB, which is otherwise a platform-neutral
session store, and none of it is reachable unless a user runs /topic.

Moved verbatim (11 methods):

  apply_telegram_topic_migration        delete_telegram_topic_binding
  enable_telegram_topic_mode            bind_telegram_topic
  disable_telegram_topic_mode           is_telegram_session_linked_to_topic
  is_telegram_topic_mode_enabled        list_unlinked_telegram_sessions_for_user
  get_telegram_topic_binding            list_telegram_topic_bindings_for_chat
  get_telegram_topic_binding_by_session

Behavior-neutral. Every body is byte-identical to what it replaced; the only
edits are the class declaration, one import, and the new module header. The
new mixin is placed first in the bases so it keeps the precedence these
methods had when they were defined on SessionDB itself; none of the three
existing mixins defines any of these names, so nothing was shadowed either
way. Per the established contract the module never imports hermes_state, and
takes _PREVIEW_RAW_SELECT and _shape_preview from hermes_state_common where
the other mixins get them.

_execute_write stays on SessionDB and resolves through the MRO unchanged.

hermes_state.py 8,604 -> 8,107 lines; the class 6,868 -> 6,370 lines and
153 -> 142 methods.
