+++ pyproject.toml
@@ -406,6 +406,7 @@ py-modules = [
   "hermes_state_portability",
   "hermes_state_schema",
   "hermes_state_search",
+  "hermes_state_telegram_topics",
   "hermes_time",
   "hermes_logging",
   "utils",
